refactor: remove pub reexport of FloatType
